requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   flask
    #   pip-tools
flask==2.0.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
itsdangerous==2.1.1
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.1.1
    # via jinja2
pep517==0.12.0
    # via pip-tools
pillow==9.0.1
    # via -r requirements.in
pip-tools==6.5.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2021.3
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via pep517
urllib3==1.26.9
    # via requests
werkzeug==2.0.3
    # via flask
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
